Many updates...

+++ public/scripts/deleteclimber.js
@@ -1,9 +1,16 @@
+$(document).ready(function() {
+    $("#climbers-link").addClass("active");
+});
+
 function deleteClimber(id) {
-    $.ajax({
-        url: '/climbers/' + id,
-        type: 'DELETE',
-        success: function(result) {
-            window.location.reload(true);
-        }
-    });
+    var confirmDelete = confirm("Are you sure you want to delete this climber?");
+    if (confirmDelete) {
+        $.ajax({
+            url: '/climbers/' + id,
+            type: 'DELETE',
+            success: function(result) {
+                window.location.reload(true);
+            }
+        });
+    }
 };
